improve: move public jwk handler to tokens file
Also remove another access function that did no authorization. Moving it to its one caller.
